Livereload triggered after tasks.
Livereload trigger added to queue as a task. 

I've noticed problem with triggering livereload before .scss compile and concat done, so changes was not visible after livereload browser refresh.
I'm newbie in Grunt and Lineman so I'm not sure is this is a good solution or is my understanding of the problem is correct. But that works for me. ;)